==> diversification/2 plurals.ipynb <==
!pip install spacy pattern
!python -m spacy download en_core_web_sm
import spacy
from pattern.en import pluralize, singularize
import pandas as pd
import random
nlp = spacy.load("en_core_web_sm")

def process_sentence(sentence):
    processed_tokens = []
    doc = nlp(sentence.strip('.'))  # remove period for processing

    for token in doc:
        word = token.text
        if token.tag_ == "NN" and singularize(word) == word and random.choice([True, False]): # pluralize only countable, singular nouns
            plural_word = pluralize(word)
            processed_tokens.append(plural_word)
        else:
            processed_tokens.append(word)

    processed_sentence = ' '.join(processed_tokens).rstrip() + '.' # no space before the period
    return processed_sentence

df = pd.read_csv('1 extended_premises_names.csv')
df['premise'] = df['premise'].apply(process_sentence)
df.to_csv('2 premises_plural.csv', index=False)
