==> diversification/6 time_money_distance.ipynb <==
!pip install spacy
!python -m spacy download en_core_web_sm
import spacy
import pandas as pd
import random
nlp = spacy.load("en_core_web_sm")

def substitute_values(sentence):
    doc = nlp(sentence)
    new_sentence = sentence

    new_values = {
        "TIME": ["a weekend", "ten days", "a week", "three weeks", "months", "two months", "six months", "years", "two years", "five years"],
        "MONEY": ["two hundred dollars", "three hundred euros", "350 pounds", "four hundred Canadian dollars", "half a thousand euros", "1000 euros", "five thousand pounds"],
        "QUANTITY": ["five miles", "two kilometers", "fifty meters", "a dozen meters", "three miles", "one kilometer", "twenty meters"],
        "GPE": ["France", "Germany", "Italy", "Portugal", "Canada", "Sweden", "Belgium"],
        "NORP": ["French", "German", "Italian", "Portuguese", "Canadian", "Swedes", "Belgians"]
    }
    for ent in reversed(list(doc.ents)):
        if ent.label_ in new_values:
            replacement_value = random.choice(new_values[ent.label_])
            new_sentence = new_sentence[:ent.start_char] + replacement_value + new_sentence[ent.end_char:]

    return new_sentence

df = pd.read_csv('5 premises_peraphrased_correct_tenses.csv')
df['2changed_premise'] = df['2changed_premise'].apply(substitute_values)
df.to_csv('6 premises_substituted_values.csv', index=False)
